Cleaned up and updated scripts
